process.py crash
Discovered whilst testing that process.py crashes in the event that a client deletes one of the currently processing tasks via the API (perhaps due to an internal timeout and subsequent cleanup).  The API prevents tasks being deleted whilst in the "running" state but not in the "completed" state.  The attempt to update the task status to "reported" in the database causes a fatal exception.